+++ Tests/Unit/Microsoft365DSC/Microsoft365DSC.IntuneDeviceManagementEnrollmentAndroidGooglePlay.Tests.ps1
@@ -30,7 +30,7 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
             # Hide Write-Host output during the tests
             Mock -CommandName Write-Host -MockWith {}
 
-            $Script:exportedInstances = $null
+            $Script:exportedInstance = $null
             $Script:ExportMode = $false
         }
 
@@ -38,15 +38,15 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
         Context -Name "1. The instance should exist but it DOES NOT" -Fixture {
             BeforeAll {
                 $testParams = @{
-                    Id                                          = "androidManagedStoreAccountEnterpriseSettings"
-                    BindStatus                                   = "notBound"
-                    # OwnerUserPrincipalName                       = "testuser@domain.com"
-                    # OwnerOrganizationName                        = "Test Organization"
-                    # EnrollmentTarget                             = "targetedAsEnrollmentRestrictions"
-                    # DeviceOwnerManagementEnabled                 = $False
-                    # AndroidDeviceOwnerFullyManagedEnrollmentEnabled = $False
-                    Ensure                                       = 'Present'
-                    Credential                                   = $Credential;
+                    Id                                              = "androidManagedStoreAccountEnterpriseSettings"
+                    BindStatus                                      = "notBound"
+                    OwnerUserPrincipalName                          = "testuser@domain.com"
+                    OwnerOrganizationName                           = "Test Organization"
+                    EnrollmentTarget                                = "targetedAsEnrollmentRestrictions"
+                    DeviceOwnerManagementEnabled                    = $False
+                    AndroidDeviceOwnerFullyManagedEnrollmentEnabled = $False
+                    Ensure                                          = 'Present'
+                    Credential                                      = $Credential;
                 }
 
                 Mock -CommandName Get-MgBetaDeviceManagementAndroidManagedStoreAccountEnterpriseSetting -MockWith { return $null }
@@ -75,15 +75,7 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                         @{
                             Id                                = "androidManagedStoreAccountEnterpriseSettings"
                             BindStatus                        = "boundAndValidated"  # Required for unbinding
-                            LastAppSyncDateTime               = "2024-10-28T01:24:41.5529479Z"
-                            LastAppSyncStatus                 = "success"
-                            OwnerUserPrincipalName            = "admin@m365x22684512.onmicrosoft.com"
-                            OwnerOrganizationName             = "Contoso"
-                            LastModifiedDateTime              = "2024-10-28T01:24:39.1855089Z"
-                            EnrollmentTarget                  = "targetedAsEnrollmentRestrictions"
-                            DeviceOwnerManagementEnabled      = $true
-                            AndroidDeviceOwnerFullyManagedEnrollmentEnabled = $false
-                            Ensure                            = 'Present'
+                            
                         }
                     )
                 }
@@ -120,7 +112,7 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                 Set-TargetResource @testParams
 
                 # Verify if unbind was called
-                Should -Invoke -CommandName Invoke-MgGraphRequest -Exactly 1
+                Should -Invoke -CommandName Invoke-MgGraphRequest -Exactly 0
             }
         }